pyproject.toml
[tool.poetry]
name = "ai-chat"
version = "0.1.0"
description = "maintain chat conversation state"
authors = ["erik aronesty <erik@q32.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ai_chat"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = {version = "^0.27.8", optional = true}
supabase = {version = "^1.0.3", optional = true}
ai-functions = "^0.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-async = "^0.1.1"
twine = ">3"
ruff = "^0.0.280"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
ignore = ['E701', 'E501']

[tool.ruff.per-file-ignores]
"__init__.py" = ['F401']


[tool.poetry.extras]
supabase = ["supabase"]
openai = ["openai"]
all = ["supabase", "openai"]
